Handle errors during database upload
