+++ lib/analyze-action.js
@@ -49,14 +49,23 @@ async function uploadDatabases(repositoryNwo, config, apiDetails, logger) {
         return;
     }
     const client = api_client_1.getApiClient(apiDetails);
-    const optInResponse = await client.request("GET /repos/:owner/:repo/code-scanning/databases", {
-        owner: repositoryNwo.owner,
-        repo: repositoryNwo.repo,
-    });
-    if (optInResponse.status !== 204) {
-        // Repository is not opted in to database uploads.
-        logger.debug("Repository is not opted in to database uploads. Skipping upload.");
-        return;
+    try {
+        await client.request("GET /repos/:owner/:repo/code-scanning/databases", {
+            owner: repositoryNwo.owner,
+            repo: repositoryNwo.repo,
+        });
+    }
+    catch (e) {
+        console.log(e);
+        if (util.isHTTPError(e)) {
+            if (e.status === 404) {
+                logger.debug("Repository is not opted in to database uploads. Skipping upload.");
+            }
+            else {
+                logger.debug(`Skipping database upload due to unknown error: ${e}`);
+            }
+            return;
+        }
     }
     const codeql = codeql_1.getCodeQL(config.codeQLCmd);
     for (const language of config.languages) {
@@ -66,17 +75,18 @@ async function uploadDatabases(repositoryNwo, config, apiDetails, logger) {
         await codeql.databaseBundle(databasePath, databaseBundlePath);
         // Upload the database bundle
         const payload = fs.readFileSync(databaseBundlePath);
-        const uploadResponse = await client.request(`PUT /repos/:owner/:repo/code-scanning/databases/${language}`, {
-            owner: repositoryNwo.owner,
-            repo: repositoryNwo.repo,
-            data: payload,
-        });
-        if (uploadResponse.status === 201) {
+        try {
+            await client.request(`PUT /repos/:owner/:repo/code-scanning/databases/${language}`, {
+                owner: repositoryNwo.owner,
+                repo: repositoryNwo.repo,
+                data: payload,
+            });
             logger.debug(`Successfully uploaded database for ${language}`);
         }
-        else {
+        catch (e) {
+            console.log(e);
             // Log a warning but don't fail the workflow
-            logger.warning(`Failed to upload database for ${language}. ${uploadResponse.data}`);
+            logger.warning(`Failed to upload database for ${language}: ${e}`);
         }
     }
 }
